+++ modules/coinswap/client/rest/grpc_query_test.go
@@ -73,8 +73,7 @@ func (s *IntegrationTestSuite) TestCoinswap() {
 	maxSupply := int64(200000000)
 	mintable := true
 	baseURL := val.APIAddress
-	uniKitty, err := coinswaptypes.GetUniDenomFromDenom(symbol)
-	s.Require().NoError(err)
+	uniKitty := coinswaptypes.GetUniDenomFromDenom(symbol)
 
 	//------test GetCmdIssueToken()-------------
 	args := []string{